some validators refactoring
