+++ pyproject.toml
@@ -50,7 +50,8 @@ include = [
 python = ">=3.8.1,<3.13"
 awscli = ">=1.33.17"
 boto3 = "^1.34.136"
-dcicutils = "^8.14.3"
+# dcicutils = "^8.14.3"
+dcicutils = "8.14.3.1b2"
 PyYAML = "^6.0.1"
 requests = "^2.31.0"
 googleapi = "^0.1.0"
