+++ submitr/submission.py
@@ -42,7 +42,7 @@ from submitr.submission_uploads import (
     lookup_ingestion_submission_from_upload_file
 )
 from submitr.utils import chars, format_path, get_health_page, is_excel_file_name, print_boxed, tobool
-from submitr.validators.utils.structured_data_validator_hook import (
+from submitr.validators.decorators import (
     define_structured_data_validator_hook,
     define_structured_data_validator_sheet_hook
 )
@@ -1924,13 +1924,10 @@ def _validate_locally(ingestion_filename: str, portal: Portal, autoadd: Optional
                                         # given (i.e. e.g. spreadsheet) object(s) into any existing ones.
                                         merge=merge,
                                         progress=None if noprogress else define_progress_callback(debug=debug),
-                                        # Doing submitted_id validation in validate_submitted_ids, in parallel.
                                         validator_hook=validator_hook,
                                         validator_sheet_hook=validator_sheet_hook,
                                         debug_sleep=debug_sleep)
     structured_data.load_file(ingestion_filename)
-    # finish_validators_hook(structured_data, valid_submission_centers=valid_submission_centers)
-    # validator_hook.finish(structured_data, valid_submission_centers=valid_submission_centers)
 
     if debug:
         PRINT("DEBUG: Finished client validation.")
